refactor: ability graph의 역량 응답에 color를 포함한다.
