+++ backend/src/main/java/wooteco/prolog/report/domain/report/abilitygraph/GraphAbilities.java
@@ -49,6 +49,7 @@ public class GraphAbilities {
             .map(ability -> new GraphAbilityDto(
                 ability.getAbilityId(),
                 ability.getName(),
+                ability.getColor(),
                 ability.getWeight(),
                 calculatePercentage(allWeight, ability),
                 ability.isPresent()
