+++ backend/src/test/resources/jsons/report_post_response.json
@@ -7,6 +7,7 @@
       {
         "id": 1,
         "name": "test parent ability1",
+        "color" : "test color1",
         "weight": 1,
         "percentage": 0.33,
         "present": true
@@ -14,6 +15,7 @@
       {
         "id": 2,
         "name": "test parent ability2",
+        "color" : "test color2",
         "weight": 1,
         "percentage": 0.33,
         "present": true
@@ -21,6 +23,7 @@
       {
         "id": 3,
         "name": "test parent ability3",
+        "color": "test color3",
         "weight": 1,
         "percentage": 0.33,
         "present": false
