+++ backend/src/test/resources/jsons/report_put_response.json
@@ -7,6 +7,7 @@
       {
         "id": 1,
         "name": "test parent ability1",
+        "color" : "test color1",
         "weight": 2,
         "percentage": 0.67,
         "present": true
@@ -14,6 +15,7 @@
       {
         "id": 3,
         "name": "test parent ability3",
+        "color": "test color3",
         "weight": 1,
         "percentage": 0.33,
         "present": false
